Create sampleData file
